Add files via upload
A perfect, efficient steering design is a key aspect of this Competition and finding one which suits this challenge was not easy and required Rigourous Testing and design and Finally we Decided to go with a steering design based on the Ackerman steering mechanism. The Ackerman steering mechanism is a design used in vehicles to ensure that the wheels turn at the correct angles when going around a corner. In this system, the front wheels are connected by a set of linkages, allowing them to turn at different angles. This is important because when a car turns, the inside wheel needs to follow a tighter curve than the outside wheel. This design improves steering efficiency and reduces tire wear by making sure all the wheels follow their natural path during a turn. It’s commonly used in cars, trucks, and other vehicles for better handling and stability.